+++ app/components/imagusu/design_system/checkbox_field_component.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module Imagusu
+  module DesignSystem
+    class CheckboxFieldComponent < FieldComponent
+    end
+  end
+end