+++ test/components/previews/imagusu/design_system/form_controls_component_preview.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+module Imagusu
+  module DesignSystem
+    class FormControlsComponentPreview < ViewComponent::Preview
+      def default
+        render_with_template
+      end
+
+      def errors
+        render_with_template
+      end
+
+      def disabled
+        render_with_template
+      end
+    end
+  end
+end