+++ test/engine_test.rb
@@ -13,5 +13,11 @@ class EngineTest < ActiveSupport::TestCase
     Rails.application.eager_load!
 
     assert_kind_of Class, Imagusu::DesignSystem::ButtonComponent
+    assert_kind_of Class, Imagusu::DesignSystem::CheckboxComponent
+    assert_kind_of Class, Imagusu::DesignSystem::FieldComponent
+    assert_kind_of Class, Imagusu::DesignSystem::RadioGroupComponent
+    assert_kind_of Class, Imagusu::DesignSystem::SelectComponent
+    assert_kind_of Class, Imagusu::DesignSystem::TextAreaComponent
+    assert_kind_of Class, Imagusu::DesignSystem::TextFieldComponent
   end
 end